No search results state for reader (#2139)
* No search results state for reader

* Adding prop types

* Removing comments

* Doing some refactoring

* Fixing styling and minor refactoring

* Fixing lint

* Making a pure component and removing a test

* Fixing css

* Fixing a test

* Reactor some code after review

* Removing unused var

* Making some more minor code improvements

* Remove extra line

* Refactoring how redux is connected

* Minor refactor

* Lint fix
