+++ app/assets/stylesheets/reader/_document_list.scss
@@ -5,6 +5,23 @@ $gray-font-color: #6b7079;
 $tag-backgrond-color: #d6d7d9;
 
 .section--document-list {
+  .section--no-search-results {
+    .cf-msg-screen-text {
+      display: none;
+    }
+
+    .cf-app-segment--alt {
+      border: none;
+      margin-top: 0;
+      padding: 4rem 4rem 2rem;
+
+      .cf-msg-screen-deck {
+        word-break: break-word;
+        margin-bottom: 0 !important;
+      }
+    }
+  }
+
   .table-icon {
     margin: 1px;
     margin-left: 5px;
